requirements.txt
Flask
flask_cors
requests
flask_mysqldb
google-auth
flask_mail
pyjwt